+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 <!-- CHANGELOG NEWEST -->
 - FIX(docs) `eventReduce: false` is the default. [#4614](https://github.com/pubkey/rxdb/pull/4614) You have to manually enable eventReduce.
+- ADD(storage-opfs) attachments support.
+- CHANGE `RxStorageInstance.getAttachmentData()` add `digest` parameter.
 
 <!-- ADD new changes here! -->